Drop sweet alert and implement own ajax delete request

+++ apps/admission/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect, HttpResponse, get_object_or_404
 from django.contrib.auth.decorators import login_required
+from django.http import JsonResponse
+from django.template.loader import render_to_string
 
 from .models import Registration, Admission
 from .forms import RegistrationForm, AdmissionForm
@@ -41,8 +43,8 @@ def registration_detail(request, registration_id):
     return render(request, 'admission/registration-detail.html', {'registration': registration})
 
 @login_required
-def edit(request, registration_id):
-    registration = get_object_or_404(Registration, id=registration_id)
+def edit_registration(request, registration_id):
+    registration = Registration.objects.get(id=registration_id)
 
     if request.POST:
         form = RegistrationForm(request.POST, instance=registration)
@@ -53,12 +55,20 @@ def edit(request, registration_id):
         form = RegistrationForm(instance=registration)
     return render(request, 'admission/edit-registration.html', {'form': form})
 
+
 @login_required
-def delete(request):
-    pk = request.POST.get('pk', 'None')
+def delete_registration(request, pk):
+    data = dict()
     registration = Registration.objects.get(pk=pk)
-    registration.delete()
-    return HttpResponse('')
+    if request.method == 'POST':
+        registration.delete()
+        return redirect('registration')
+    else:
+        context ={'registration': registration}
+        data['html_form'] = render_to_string('admission/partial-registration-delete.html', context, request=request)
+        return JsonResponse(data)
+
+
 
 @login_required
 def admission(request):
